moved xpath functions to separate package
